LPD-25944 Refactor Selection filters
